requirements.txt
Django==3.1.14
django-cors-headers==3.9.0
django-simple-history==3.0.0
django-storages==1.12.3
djangorestframework==3.11.2
drf-yasg==1.20.0
psycopg2==2.9.1
psycopg2-binary==2.9.1